// pms-ui/src/app/faculty/drives/components/types.ts
export interface Company {
    _id: string;
    name: string;
    site?: string;
    branch: string;
    desc?: string;
    email?: string;
    ph_no?: string;
    avg_salary?: number;
    placed_students?: string[];
}

export interface Job {
    _id: string;
    company: string;
    drive: string;
    title: string;
    desc?: string;
    loc?: string;
    requirement?: string;
    experience?: number;
    salary?: number;
    join_date: string | Date;
    last_date: string | Date;
    contact_person?: string;
    contact_email?: string;
    additional_instructions?: string;
    form_link?: string;
    eligible_students?: string[];
    applied_students?: string[];
    selected_students?: string[];
    stage_students?: string[][];
}